pyproject.toml
# app_stremlit/pyproject.toml
[project]
name = "app-stremlit"
version = "0.1.0"
dependencies = [
    "streamlit==1.37.1",
    "segment-geospatial==0.12.0",
    "groundingdino-py==0.4.0",
    "transformers==4.55.0",
    "supervision==0.6.0",
    "timm==1.0.19",
    "opencv-python-headless==4.12.0.88",
    "rasterio==1.4.3",
    "fiona==1.10.1",
    "pyproj==3.7.1",
    "geopandas==1.1.1",
    "shapely==2.1.1",
    "folium==0.20.0",
    "rio-tiler==7.8.1",
    "rio-cogeo==5.4.2",
    "python-dotenv==1.0.1",
    "pandas==2.2.2",
    "altair==5.3.0",
]
requires-python = ">=3.8"
packages = [
    { include = "pipeline" },
    { include = "samgeo_stub" },
    { include = "nl_query" },
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"